Serialise terminal IO
If you defined more than one scrobbling account, then olscrbl does the
submissions in parallel via ‘n-par-for-each’, which uses threads to do
the parallel processing. This works mostly seamlessly, because there is
no writable data shared at all, except for the terminal, which is used
to output progress messages to.

This is largely a cosmetical problem, but since it's rather easy to
fix...

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>